// frontend/src/app/api/model/headDim.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * llmdex
 * OpenAPI spec version: 0.1.0
 */
import type { HeadDimMismatch } from './headDimMismatch';
import type { HeadDimSource } from './headDimSource';

export interface HeadDim {
  value?: number | null;
  source?: HeadDimSource;
  mismatch?: HeadDimMismatch | null;
}